requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
tqdm
sentence-transformers
faiss-cpu
numpy
openai
python-dotenv